More clean up.
 * Fixed --grouping-as-class option
 * Added this option + docgen to CI
