+++ profiles/openconfig/openconfig-telem.json
@@ -9,6 +9,7 @@
                 "url": "http://github.com/openconfig/public.git",
 		"commits": [
 		    {
+		        "commitid": "4ab22755c2484e893b2c2b6d88f800b8dd41bb09",
 			"file": [
 		      	    "release/models/openconfig-extensions.yang",
 		      	    "release/models/telemetry/openconfig-telemetry.yang"
@@ -20,6 +21,7 @@
                 "url": "http://github.com/YangModels/yang.git",
 		"commits": [
 		    {
+		        "commitid": "1dc8f5417bca052bb2248d077b845581056f1036",
 			"file": [
 			    "standard/ietf/RFC/ietf-inet-types.yang"
 			]
