+++ ydkgen/builder/__init__.py
@@ -15,4 +15,5 @@
 # ------------------------------------------------------------------
 
 from ._api_model_builder import ApiModelBuilder, GroupingClassApiModelBuilder
+from ._pyang_model_builder import PyangModelBuilder
 from ._types_extractor import TypesExtractor
